auth/requirements.txt
fastapi==0.115.0
uvicorn==0.32.1
pydantic==2.10.3
pydantic-settings==2.7.0
flake8==7.1.1
SQLAlchemy==2.0.37
Werkzeug==3.1.3
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0